requirements.txt
feedparser
spacy
schedule
requests
beautifulsoup4
mysql-connector-python
pyyaml
google-cloud-language

